requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.2.0
Django==3.2.5
django-extensions==3.2.3
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
drf-extensions==0.7.1
drf-nested-routers==0.93.4
idna==3.4
install==1.3.5
PyJWT==2.8.0
pytz==2023.3
PyYAML==6.0.1
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.4